Reorganized common components
