+++ auction/src/components/Common/BidButton.js
@@ -4,7 +4,7 @@ import { useHistory } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faChevronRight, faGavel } from '@fortawesome/free-solid-svg-icons';
 
-import Button from '../Controls/Button';
+import Button from './Button';
 
 const BidButton = ({ text, hammer, product, onClick, ...props }) => {
     const history = useHistory();
