+++ auction/src/components/Common/Container/index.js
@@ -1,0 +1,1 @@
+export { default } from './Container';
