+++ auction/src/components/Common/Divider/index.js
@@ -1,0 +1,1 @@
+export { default } from './Divider';
