+++ auction/src/components/Common/FormGroup/index.js
@@ -1,0 +1,1 @@
+export { default } from './FormGroup';
