+++ auction/src/components/Common/Label/Label.js
