+++ auction/src/components/Common/Label/index.js
@@ -1,0 +1,1 @@
+export { default } from './Label';
