+++ auction/src/components/Common/Social/index.js
@@ -1,0 +1,1 @@
+export { default } from './Social';
