+++ auction/src/components/Common/TextField/index.js
@@ -1,0 +1,1 @@
+export { default } from './TextField';
