+++ auction/src/components/Common/WishlistButton.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faHeart } from '@fortawesome/free-solid-svg-icons';
 
-import Button from '../Controls/Button';
+import Button from './Button';
 
 const WishlistButton = () => {
     return (
