+++ auction/src/components/PageLayout/NavBar/SearchBar/index.js
@@ -1,0 +1,1 @@
+export { default } from './SearchBar';
