+++ auction/src/components/PageLayout/PageLayout.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import TopBar from './Common/TopBar';
-import NavBar from './Common/NavBar';
+import TopBar from './TopBar';
+import NavBar from './NavBar';
 
 const PageLayout = ({ children }) => {
     return (
