+++ auction/src/components/PageLayout/TopBar/index.js
@@ -1,0 +1,1 @@
+export { default } from './TopBar';
