+++ auction/src/components/ProductPage/ProductDetails/ProductDetails.js
@@ -1,11 +1,9 @@
 import React, { useState, useEffect } from 'react';
-
-import TextField from '../../Controls/TextField';
-import FormGroup from '../../Controls/FormGroup';
-import WishlistButton from '../../Common/WishlistButton';
 import axios from 'axios';
+
+import { TextField, FormGroup, WishlistButton, BidButton } from '../../Common';
+
 import config from '../../../config';
-import BidButton from '../../Common/BidButton';
 
 const getBids = async (productID, setBids) => {
     try {
